[Fix] Zombie Animation Fixed
1. 애니메이션 반복되던거 오류 수정
