--- Assets/2.Model/Prefabs/ZombieTest.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 762f43ad230335b499158f1fd7e0a0eb
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
