+++ Assets/3.Script/Zombie/ZombieController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5925bf1189dbc834e890052e005b7342
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
